home-quarantine/requirements.txt
requests==2.20.1
google-api-python-client==1.7.11
